refactor: rename mod `utils` to `u256`
